refactor(frontend): move the route table out of App.tsx
Importing the table for the breadcrumb coverage test also ran
createBrowserRouter and pulled the lazy page graph, which is a routing side
effect in a unit test — and the export needed an eslint-disable to survive the
react-refresh rule.

routes.tsx holds the table and constructs nothing; App.tsx keeps the QueryClient
and createBrowserRouter(routes). The disable is deleted rather than relocated,
which took one more step: the rule then fired from the other side, since the new
file exported a non-component while defining FeedbackRedirect. That redirect now
has its own file, following the flat-component precedent of PrototypeRenderer.

No eslint-disable remains in this branch.
